% $Id: gstat-internal.Rd,v 1.5 2007-06-08 06:45:51 edzer Exp $
\name{gstat-internal}
\alias{vgm.panel.xyplot}
\alias{load.variogram.model}
\alias{gstat.formula}
\alias{gstat.formula.predict}
\alias{gstat.debug}
\alias{gstat.set}
\alias{cross.name}
\alias{gstat.load.set}
\alias{panel.pointPairs}

\title{Gstat Internal Functions}
\description{gstat internal functions}
\author{Edzer J. Pebesma}
\note{these functions should not be called by users directly}
\keyword{internal}
