\name{gstat-internal}
\alias{vgm.panel.xyplot}
\alias{xvgm.panel.xyplot}
\alias{vgm.dir.panel.xyplot}
\alias{load.variogram.model}
\alias{gstat.formula}
\alias{gstat.formula.predict}
\alias{gstat.debug}
\alias{gstat.set}
\alias{gstat.load.set}
\alias{panel.point.pairs}

\title{Gstat Internal Functions}
\description{gstat internal functions}
\author{Edzer J. Pebesma}
\note{these functions should not be called by users directly}
\keyword{internal}
