% $Id: gstat-internal.Rd,v 1.6 2008-03-10 10:36:04 edzer Exp $
\name{gstat-internal}
\alias{load.variogram.model}
\alias{gstat.formula}
\alias{gstat.formula.predict}
\alias{gstat.debug}
\alias{gstat.set}
\alias{cross.name}
\alias{gstat.load.set}

\title{Gstat Internal Functions}
\description{gstat internal functions}
\author{Edzer Pebesma}
\note{these functions are not meant to be called by users directly}
\keyword{internal}
